# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SindÂncora API",
        "description": "API REST do SindÂncora — SaaS de Gestão Condominial",
        "contact": {
            "email": "suporte@sindancora.com.br"
        },
        "license": {
            "name": "Proprietário"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Base"
        }
    ],
    "paths": {
        "/v1/app/announcements": {
            "get": {
                "tags": [
                    "App"
                ],
                "summary": "Listar comunicados (filtros: search, status, condominium_id)",
                "operationId": "appAnnouncementsIndex",
                "responses": {
                    "200": {
                        "description": "Lista paginada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "App"
                ],
                "summary": "Criar comunicado (action: draft|publish; anexos opcionais)",
                "operationId": "appAnnouncementsStore",
                "responses": {
                    "201": {
                        "description": "Comunicado criado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/app/announcements/{id}": {
            "get": {
                "tags": [
                    "App"
                ],
                "summary": "Detalhe do comunicado com anexos",
                "operationId": "appAnnouncementsShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comunicado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/app/announcements/{id}/publish": {
            "post": {
                "tags": [
                    "App"
                ],
                "summary": "Publicar comunicado (rascunho ou agendado)",
                "operationId": "appAnnouncementsPublish",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comunicado publicado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/v1/app/charges": {
            "get": {
                "tags": [
                    "App"
                ],
                "summary": "Listar cobranças com KPIs (filtros: condominium_id, status, reference_month, search)",
                "operationId": "appChargesIndex",
                "responses": {
                    "200": {
                        "description": "Lista paginada + KPIs"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/app/charges/{id}": {
            "get": {
                "tags": [
                    "App"
                ],
                "summary": "Detalhe da cobrança",
                "operationId": "appChargesShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cobrança"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/app/expenses": {
            "get": {
                "tags": [
                    "App"
                ],
                "summary": "Listar contas a pagar (filtros: condominium_id, status, from, to)",
                "operationId": "appExpensesIndex",
                "responses": {
                    "200": {
                        "description": "Lista paginada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            },
            "apiKey": {
                "type": "http",
                "description": "API Key do tenant no formato sk_live_... enviada como Bearer token.",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}